requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
cryptography==1.9
cssselect==1.0.1
enum34==1.1.6
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
ipaddress==1.0.18
lxml==3.8.0
numpy==1.13.1
pandas==0.20.2
parsel==1.2.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.1.0
python-dateutil==2.6.0
pytz==2017.2
queuelib==1.4.2
requests==2.18.1
Scrapy==1.4.0
service-identity==17.0.0
six==1.10.0
tabula-py==0.9.0
Twisted==17.5.0
urllib3==1.21.1
w3lib==1.17.0
wheel==0.24.0
zope.interface==4.4.2
